Refactor location helper into dedicated services
