+++ test/Unit/Clusterer/AnniversaryClusterStrategyTest.php
@@ -27,7 +27,7 @@ final class AnniversaryClusterStrategyTest extends TestCase
     public function keepsHighestScoringAnniversariesWithinLimit(): void
     {
         $strategy = new AnniversaryClusterStrategy(
-            new LocationHelper(),
+            LocationHelper::createDefault(),
             minItemsPerAnniversary: 3,
             minDistinctYears: 2,
             maxClusters: 1,
@@ -75,7 +75,7 @@ final class AnniversaryClusterStrategyTest extends TestCase
     public function returnsEmptyWhenAnniversaryLacksDistinctYears(): void
     {
         $strategy = new AnniversaryClusterStrategy(
-            new LocationHelper(),
+            LocationHelper::createDefault(),
             minItemsPerAnniversary: 3,
             minDistinctYears: 3,
             maxClusters: 0,
@@ -96,7 +96,7 @@ final class AnniversaryClusterStrategyTest extends TestCase
     public function skipsGroupsBelowMinimumItemCount(): void
     {
         $strategy = new AnniversaryClusterStrategy(
-            new LocationHelper(),
+            LocationHelper::createDefault(),
             minItemsPerAnniversary: 4,
             minDistinctYears: 2,
             maxClusters: 0,
